refactor: StringifyMarkdownOptionsとVFMSettingsの共通部分をSerializablePluginOptionsに抽出
